scripts/requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
picamera==1.13
pkg-resources==0.0.0
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyserial==3.5
requests==2.25.0
urllib3==1.26.2
